chore: Renamed Core/Singletontmp.mts to Core/singleton.mts.
