+++ packages/architectura/src/Core/index.mts
@@ -9,6 +9,6 @@ export * from "./rich-client-request.mjs";
 export * from "./rich-server-response.mjs";
 export * from "./Server.mjs";
 export * from "./Server/index.mjs";
-export * from "./Singletontmp.mjs";
+export * from "./singleton.mjs";
 export * from "./Time.mjs";
 export * from "./Time/index.mjs";
