+++ packages/architectura/src/Core/singleton.mts
